Fixet fejl og tilføjet.

+++ client.lua
@@ -1,17 +1,22 @@
-cooldownactive = false
-bilcooldownactive = false
-pakketaget = false
-annulervar = true
-paabegyndvar = false
-parkerbilvar = true
-spawnbilvar = false
-player = GetPlayerPed(-1) 
-biltargetvar = false
+local cooldownactive = false
+local bilcooldownactive = false
+local pakketaget = false
+local annulervar = true
+local paabegyndvar = false
+local parkerbilvar = true
+local spawnbilvar = false
+local biltargetvar = false
+local player = PlayerPedId() -- Changed from GetPlayerPed(-1)
+local biltargetvar = false
+local person = nil
+local bil = nil
+local modtager = nil
+local prop = nil
+local randomX, randomY, randomZ, randomH = 0, 0, 0, 0
+local playergotocoords = vector3(0, 0, 0)
+local BilNetId = nil
 
-vRP = Proxy.getInterface("vRP")
-vRPclient = Tunnel.getInterface("vRP","Vrp-postbud")
-
-HT = nil
+local HT = nil
 
 Citizen.CreateThread(function()
     while HT == nil do
@@ -20,174 +25,174 @@ Citizen.CreateThread(function()
     end
 end)
 
+Citizen.CreateThread(function()
+    if not DoesEntityExist(person) then
+        RequestModel(Config.person)
+        while not HasModelLoaded(Config.person) do
+            Citizen.Wait(100)
+        end
 
-if DoesEntityExist(person) == false then
-  RequestModel(Config.person)
-  while not HasModelLoaded(Config.person) do
-      Citizen.Wait(100)
-  end
-
-  person = CreatePed(1, GetHashKey(Config.person), Config.pedCoords, true, true)
-  FreezeEntityPosition(person, true)
-  SetEntityInvincible(person, true)
-  SetBlockingOfNonTemporaryEvents(person, true)
-  FreezeEntityPosition(
-    person --[[ Entity ]], 
-    true --[[ boolean ]]
-  )
-  local pedNetId = NetworkGetNetworkIdFromEntity(person)
-  exports.ox_target:addEntity(pedNetId, Config.pedoptions)
-end
-
-
-
-
-RegisterCommand("postmenu", function()
-  HT.TriggerServerCallback('Postbud:hasgruop', function(result)
-    if result then
-      postmenu()
-    else 
-      lib.notify({
-        title = 'fejl',
-        description = 'Du er ikke postmand',
-        type = 'error'
-      })
+        person = CreatePed(1, GetHashKey(Config.person), Config.pedCoords, true, true)
+        FreezeEntityPosition(person, true)
+        SetEntityInvincible(person, true)
+        SetBlockingOfNonTemporaryEvents(person, true)
+        FreezeEntityPosition(person, true)
+        local pedNetId = NetworkGetNetworkIdFromEntity(person)
+        exports.ox_target:addEntity(pedNetId, Config.pedoptions)
     end
-end, Config.gruop)    
 end)
 
--- Ox Menu -- 
+RegisterCommand("postmenu", function()
+    HT.TriggerServerCallback('Postbud:hasgruop', function(result)
+        if result then
+            postmenu()
+        else
+            notifyerror()
+        end
+    end, Config.gruop)
+end)
 
 function postmenu()
-  lib.registerContext({
-    id = 'post:menu',
-    title = 'Post Menu📦',
-    options = {
-      {
-        title = 'Oversigt',
-      },
-      {
-        title = 'Spawn køretøj',
-        description = 'her kan du spawne dit køretøj',
-        icon = 'car',
-        event = 'postbil',
-        disabled = spawnbilvar
-      },
-      {
-        title = 'parker køretøj',
-        description = 'her kan du parkere dit køretøj',
-        icon = 'car',
-        event = 'postbilparker',
-        disabled = parkerbilvar,
-      },
-      {
-        title = 'Påbegynd opgave',
-        description = 'Gå igang med en opgave',
-        icon = 'envelope',
-        event = 'startopgave',
-        disabled = paabegyndvar
-      },
-      {
-        title = 'Annuler Opgave',
-        description = 'Stop din igangværende opgave',
-        icon = 'xmark',
-        event = 'stopopgave',
-        disabled = annulervar
-      },
-    }
-  })
-  lib.showContext('post:menu')
+    lib.registerContext({
+        id = 'post:menu',
+        title = 'Post Menu📦',
+        options = {
+            {
+                title = 'Postbud - Oversigt',
+            },
+            {
+                title = 'Spawn køretøj',
+                description = 'Her kan du hente dit køretøj.',
+                icon = 'car',
+                event = 'postbil',
+                disabled = spawnbilvar
+            },
+            {
+                title = 'Parker køretøj',
+                description = 'Du skal være henne ved posthuset for at parkere bilen.',
+                icon = 'car',
+                event = 'postbilparker',
+                disabled = parkerbilvar,
+            },
+            {
+                title = 'Påbegynd opgave',
+                description = 'Start en opgave.',
+                icon = 'envelope',
+                event = 'startopgave',
+                disabled = paabegyndvar
+            },
+            {
+                title = 'Annuller Opgave',
+                description = 'Stop din nuværende opgave.',
+                icon = 'xmark',
+                event = 'stopopgave',
+                disabled = annulervar
+            },
+        }
+    })
+    lib.showContext('post:menu')
 end
-
---------------
--- info center
---------------
 
 RegisterNetEvent("openinfocenter")
 AddEventHandler("openinfocenter", function()
-    if ESX.PlayerData.job.name == 'postmand' then
-      local info = lib.alertDialog({
-        header = "hej",
-        content = 'som postmand er dit job at levere pakker til byens borgere skriv /postmenu for at komme igang',
-        centered = true,
-        cancel = true
-    })
-    else
-      lib.notify({
-        title = 'fejl',
-        description = 'Du er ikke postmand',
-        type = 'error'
-    })
-    end
+    HT.TriggerServerCallback('Postbud:hasgruop', function(result)
+        if result then
+            local info = lib.alertDialog({
+                header = "Postbud - Chef",
+                content = 'Som postbud er dit job at levere pakker til byens borgere brug "/postmenu" for at komme igang.',
+                centered = true,
+                cancel = true
+            })
+        else
+            notifyerror()
+        end
+    end, Config.gruop)
 end)
-
-----------
--- postbil
-----------
 
 RegisterNetEvent("postbil")
 AddEventHandler("postbil", function()
-  local playercoords = GetEntityCoords(player)
-  local forskel = GetDistanceBetweenCoords(playercoords, Config.bilspawn, true)
+    local playercoords = GetEntityCoords(player)
+    local forskel = #(playercoords - Config.bilspawn)
   
-  if bilcooldownactive == false then
-      if forskel <= Config.bilspawndistance then
-          RequestModel(Config.bilnavn)
-          while not HasModelLoaded(Config.bilnavn) do
-              Citizen.Wait(500)
-          end
+    if not bilcooldownactive then
+        if forskel <= Config.bilspawndistance then
+            RequestModel(Config.bilnavn)
+            while not HasModelLoaded(Config.bilnavn) do
+                Citizen.Wait(500)
+            end
   
-          bil = CreateVehicle(GetHashKey(Config.bilnavn), Config.bilspawn, true, false)
-          TriggerEvent("bilcooldown")
-          spawnbilvar = true
-          parkerbilvar = false
-          BilNetId = NetworkGetNetworkIdFromEntity(bil)
-      else
-          lib.notify({
-              title = 'fejl',
-              description = 'Du er for langt væk',
-              type = 'error'
-          })
-      end
-  else
-      lib.notify({
-          title = 'fejl',
-          description = 'Du skal vente lidt før at du kan spawne en ny bil',
-          type = 'error'
-      })
-  end  
+            bil = CreateVehicle(GetHashKey(Config.bilnavn), Config.bilspawn, true, false)
+            carcooldown()
+            spawnbilvar = true
+            parkerbilvar = false
+            BilNetId = NetworkGetNetworkIdFromEntity(bil)
+        else
+            lib.notify({
+                title = 'Postbud',
+                description = 'Du er for langt væk.',
+                type = 'error'
+            })
+        end
+    else
+        lib.notify({
+            title = 'Postbud',
+            description = 'Du skal vente lidt før at du kan hente en ny bil.',
+            type = 'error'
+        })
+    end  
 end)
 
-
--- parker function
 RegisterNetEvent("postbilparker")
 AddEventHandler("postbilparker", function()
-    DeleteEntity(bil)
-    spawnbilvar = false
-    parkerbilvar = true
+    local playerPed = PlayerPedId()
+    local playerCoords = GetEntityCoords(playerPed)
+    local distance = GetDistanceBetweenCoords(playerCoords, Config.bilspawn.x, Config.bilspawn.y, Config.bilspawn.z, true)
+
+    if distance <= Config.maxDistanceToRemoveCar then
+        if IsPedInVehicle(playerPed, bil, false) then
+            TaskLeaveVehicle(playerPed, bil, 0)
+            Citizen.Wait(1000) 
+        end
+
+        DoScreenFadeOut(1000) 
+        Citizen.Wait(1000) 
+
+        if DoesEntityExist(bil) then
+            DeleteEntity(bil)
+            spawnbilvar = false
+            parkerbilvar = true
+        end
+
+        Citizen.Wait(1000)
+        DoScreenFadeIn(1000) 
+    else
+        lib.notify({
+            title = 'Postbud',
+            description = 'Du er for langt væk til at fjerne bilen.',
+            type = 'error'
+        })
+    end
 end)
 
--- bil cooldown function
-RegisterNetEvent("bilcooldown")
-AddEventHandler("bilcooldown", function()
+
+
+function carcooldown()
     bilcooldownactive = true
-    Wait(Config.cooldownbil)
+    Citizen.Wait(Config.cooldownbil)
     bilcooldownactive = false
-end)
+end
 
--- tag pakke ud af bil function
 RegisterNetEvent("pakkeud")
 AddEventHandler("pakkeud", function()
     pakketaget = true
-    local ped = GetPlayerPed(-1)
-    local boneIndex = GetPedBoneIndex(ped, 60309)
+    local boneIndex = GetPedBoneIndex(player, 60309)
 
     RequestAnimDict("anim@heists@box_carry@")
     while not HasAnimDictLoaded("anim@heists@box_carry@") do
         Citizen.Wait(100)
     end
 
-    TaskPlayAnim(ped, "anim@heists@box_carry@", "idle", 8.0, -8, -1, 49, 0, false, false, false)
+    TaskPlayAnim(player, "anim@heists@box_carry@", "idle", 8.0, -8, -1, 49, 0, false, false, false)
 
     Citizen.Wait(500)
 
@@ -198,44 +203,34 @@ AddEventHandler("pakkeud", function()
     end
 
     prop = CreateObject(modelHash, 0.0, 0.0, 0.0, true, true, false)
-    AttachEntityToEntity(prop, ped, boneIndex, 0.025, 0.08, 0.255, -145.0, 290.0, 0.0, true, true, false, true, 1, true)
-    local xRotation = 0.0
-    local yRotation = 0.0
-    local zRotation = 90.0
+    AttachEntityToEntity(prop, player, boneIndex, 0.025, 0.08, 0.255, -145.0, 290.0, 0.0, true, true, false, true, 1, true)
+    local xRotation, yRotation, zRotation = 0.0, 0.0, 90.0
 
     SetEntityRotation(prop, xRotation, yRotation, zRotation, 2, true)
 end)
 
--- remove target
 RegisterNetEvent("respawncar")
 AddEventHandler("respawncar", function()
-  DeleteVehicle(bil)
+    if DoesEntityExist(bil) then
+        DeleteVehicle(bil)
 
-    RequestModel(Config.bilnavn)
-          while not HasModelLoaded(Config.bilnavn) do
-              Citizen.Wait(500)
-          end
+        RequestModel(Config.bilnavn)
+        while not HasModelLoaded(Config.bilnavn) do
+            Citizen.Wait(500)
+        end
 
-    bil = CreateVehicle(GetHashKey(Config.bilnavn), carspawn, true, false)
+        bil = CreateVehicle(GetHashKey(Config.bilnavn), Config.carspawn, true, false)
+    end
 end)
 
----------------
--- post opgaver
----------------
-
--- opgave start
 RegisterNetEvent("startopgave")
-
 AddEventHandler("startopgave", function()
     if DoesEntityExist(bil) then
-        if cooldownactive == false then
+        if not cooldownactive then
             local totalCoords = #Config.coords
             local randomIndex = math.random(1, totalCoords)
             randomCoords = Config.coords[randomIndex]
-            randomX = randomCoords[1]
-            randomY = randomCoords[2]
-            randomZ = randomCoords[3]
-            randomH = randomCoords[4]
+            randomX, randomY, randomZ, randomH = randomCoords[1], randomCoords[2], randomCoords[3], randomCoords[4]
             SetNewWaypoint(randomX, randomY)
 
             exports.ox_target:addEntity(BilNetId, Config.pakkeoptions)
@@ -249,7 +244,7 @@ AddEventHandler("startopgave", function()
 
             lib.notify({
                 title = 'Du startede en opgave',
-                description = 'Kør hen til huset på din gps og aflever pakken',
+                description = 'Kør hen til huset på din gps og aflever pakken.',
                 type = 'info'
             })
 
@@ -263,7 +258,7 @@ AddEventHandler("startopgave", function()
             FreezeEntityPosition(modtager, true)
             SetEntityInvincible(modtager, true)
             SetBlockingOfNonTemporaryEvents(modtager, true)
-            TaskStandStill(modtager, 100000000000000000000000000000)
+            FreezeEntityPosition(modtager, true)
 
             playergotocoords = GetOffsetFromEntityInWorldCoords(modtager, 0.0, 1.0, 0.0)
             local ModtagerNetId = NetworkGetNetworkIdFromEntity(modtager)
@@ -271,46 +266,41 @@ AddEventHandler("startopgave", function()
             exports.ox_target:addEntity(ModtagerNetId, Config.modtageroptions)
         else
             lib.notify({
-                title = 'fejl',
-                description = 'du skal vente før at du kan starte endu en opgave',
+                title = 'Postbud',
+                description = 'Du skal vente før at du kan starte endu en opgave',
                 type = 'error'
             })
         end
     else
         lib.notify({
-            title = 'fejl',
+            title = 'Postbud',
             description = 'du skal have en postbil for at starte en opgave',
             type = 'error'
         })
     end
 end)
 
--- opgave færdig
 RegisterNetEvent("opgavefærdig")
 AddEventHandler("opgavefærdig", function()
-    carspawn = GetEntityCoords(bil)
-    local boneIndex2 = GetPedBoneIndex(GetPlayerPed(-1), 57005)
-    local boneIndex3 = GetPedBoneIndex(modtager, 57005)
     local oppositeHeading = (randomH + 180) % 360
-    local player = GetPlayerPed(-1)
     TaskGoStraightToCoord(player, playergotocoords, 1, -1, oppositeHeading, 0.1)
-    SetEntityHeading(GetPlayerPed(-1), oppositeHeading)
+    SetEntityHeading(player, oppositeHeading)
 
     Wait(2000)
 
     RequestAnimDict("mp_common")
-  while not HasAnimDictLoaded("mp_common") do
-      Citizen.Wait(100)
-  end
+    while not HasAnimDictLoaded("mp_common") do
+        Citizen.Wait(100)
+    end
 
-  DetachEntity(prop, true, true)
-  AttachEntityToEntity(prop, GetPlayerPed(-1), boneIndex2, 0.025, 0.08, 0.255, -145.0, 290.0, 0.0, true, true, false, true, 1, true)
+    DetachEntity(prop, true, true)
+    AttachEntityToEntity(prop, player, boneIndex2, 0.025, 0.08, 0.255, -145.0, 290.0, 0.0, true, true, false, true, 1, true)
 
-  ClearPedTasks(GetPlayerPed(-1))
-  ClearPedTasks(modtager)
+    ClearPedTasks(player)
+    ClearPedTasks(modtager)
 
-  TaskPlayAnim(GetPlayerPed(-1), "mp_common", "givetake1_a", 8.0, -8, -1, 0, 0, false, false, false)
-  TaskPlayAnim(modtager, "mp_common", "givetake1_b", 8.0, -8, -1, 0, 0, false, false, false)
+    TaskPlayAnim(player, "mp_common", "givetake1_a", 8.0, -8, -1, 0, 0, false, false, false)
+    TaskPlayAnim(modtager, "mp_common", "givetake1_b", 8.0, -8, -1, 0, 0, false, false, false)
     Citizen.Wait(1000)
     DetachEntity(prop, true, true)
     AttachEntityToEntity(prop, modtager, boneIndex3, 0.025, 0.08, 0.255, -145.0, 290.0, 0.0, true, true, false, true, 1, true)
@@ -329,38 +319,40 @@ AddEventHandler("opgavefærdig", function()
     lib.notify({title = 'Færdig', description = 'Du færdig gjorde din opgave', type = 'succes'})
 end)
 
--- stop opgave
 RegisterNetEvent("stopopgave")
 AddEventHandler("stopopgave", function()
     biltargetvar = false
     local alert = lib.alertDialog({
-    header = 'Stop opgave',
-    content = 'Vil du gerne stoppe din nuværende opgave',
-    centered = true,
-    cancel = true
+        header = 'Stop opgave',
+        content = 'Vil du gerne stoppe din nuværende opgave',
+        centered = true,
+        cancel = true
     })
     if alert == "confirm" then
-    parkerbilvar = false
-    annulervar = true
-    paabegyndvar = false
-    DeletePed(modtager)
-    DeleteWaypoint()
-    DeleteEntity(prop)
-    lib.notify({title = 'Stop', description = 'Du stoppede din opgave', type = 'info'})
-    TriggerEvent("respawncar")
+        parkerbilvar = false
+        annulervar = true
+        paabegyndvar = false
+        DeletePed(modtager)
+        DeleteWaypoint()
+        DeleteEntity(prop)
+        lib.notify({title = 'Stop', description = 'Du stoppede din opgave', type = 'info'})
+        TriggerEvent("respawncar")
     else
-      lib.notify({title = 'Info', description = 'Du fortsatte med din opgave', type = 'info'})
+        lib.notify({title = 'Info', description = 'Du fortsatte med din opgave', type = 'info'})
     end
 end)
 
--- Opgave cooldown
 RegisterNetEvent("cooldown")
 AddEventHandler("cooldown", function()
     cooldownactive = true
-    Wait(Config.cooldown)
+    Citizen.Wait(Config.cooldown)
     cooldownactive = false
 end)
 
-RegisterCommand("test", function()
-  TriggerServerEvent("betaling")
-end)
+function notifyerror()
+    lib.notify({
+        title = 'Postbud',
+        description = 'Du er ikke postmand',
+        type = 'error'
+    })
+end
